9 time table

+++ 0x02-functions_nested_loops/9-times_table.c
@@ -5,7 +5,7 @@
  */
 void times_table(void)
 {
-	int i, j;
+	float i, j;
 
 	/**
 	 * for i start from 0 to 9
